Migrate presentational-component to TS
